claude-agent-sdk-tutorial/requirements.txt
# Core dependencies for Databricks Apps deployment
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
claude-agent-sdk>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
